backend/pyproject.toml
[tool.poetry]
name = "quadrant-backend"
version = "0.1.0"
description = "Backend services for the Quadrant mobile application"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.32" }
asyncpg = "^0.29.0"
alembic = "^1.13.3"
pydantic-settings = "^2.4.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
redis = "^5.0.8"
httpx = "^0.27.0"
celery = "^5.4.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"
httpx = { version = "^0.27.0", extras = ["cli"] }
ruff = "^0.6.3"
black = "^24.8.0"
mypy = "^1.11.1"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "DTZ", "ASYNC"]
ignore = ["E203", "E266", "E501"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
